時間表示のメソッドを共通化
