+++ client/src/components/organisms/ActiveTask.js
@@ -6,6 +6,7 @@ import styled from 'styled-components';
 
 import { TASK_STATUS } from '../../const';
 import { activeTaskModule } from '../../modules/activeTaskModule';
+import { milliSecToHhmmssms } from '../../utils/taskUtil';
 import Button from '../atoms/Button';
 
 const propTypes = {
@@ -42,21 +43,6 @@ const ActiveTask = props => {
       clearInterval(timer);
     });
     props.updateState({ timers: [] });
-  };
-
-  const milliSecToHhmmssms = milliSeconds => {
-    const hour = Math.floor(milliSeconds / (60 * 60 * 1000));
-    const minute = Math.floor((milliSeconds % (60 * 60 * 1000)) / (60 * 1000));
-    const second = Math.floor((milliSeconds % (60 * 1000)) / 1000);
-    const milliSecond = Math.floor(milliSeconds % 1000);
-
-    const hh = ('00' + hour).slice(-2);
-    const mm = ('00' + minute).slice(-2);
-    const ss = ('00' + second).slice(-2);
-    const ms = ('000' + milliSecond).slice(-3, -2);
-
-    const formatedTime = `${hh}:${mm}:${ss}.${ms}`;
-    return formatedTime;
   };
 
   useEffect(() => {
