+++ client/webpack.common.js
@@ -32,7 +32,7 @@ module.exports = {
         ],
       },
       {
-        test: /\.tsx$/,
+        test: /\.tsx?$/,
         use: 'ts-loader',
       },
     ],
